Revert "Add T3 Connect onboarding for mobile and web (#3765)"
This reverts commit fa6073e2e13b6a41cfd6c3834c300408e9d1d6d8.
