+++ apps/mobile/src/features/cloud/CloudAuthProvider.tsx
@@ -18,7 +18,6 @@ import {
   setAgentAwarenessRelayTokenProvider,
   unregisterAgentAwarenessDeviceForCurrentUser,
 } from "../agent-awareness/remoteRegistration";
-import { requestConnectOnboarding } from "./connectOnboarding";
 import { resolveCloudPublicConfig, resolveRelayClerkTokenOptions } from "./publicConfig";
 
 function resetManagedRelayTokenCache() {
@@ -67,14 +66,6 @@ function CloudAuthBridge(props: { readonly children: ReactNode }) {
     const previousObservedAccount = observedAccountRef.current;
     const nextAccount = isSignedIn && userId ? userId : null;
     observedAccountRef.current = nextAccount;
-
-    // Every sign-in that completes during this session (a cold start observes
-    // undefined → account and must not re-prompt) requests the T3 Connect
-    // onboarding sheet — sign-out clears the connected environments, so each
-    // new session starts with no devices to reach.
-    if (previousObservedAccount === null && nextAccount !== null) {
-      requestConnectOnboarding(nextAccount);
-    }
 
     const queueAccountCleanup = (
       previous: {
