+++ apps/mobile/src/features/connection/ConnectionEnvironmentRow.tsx
@@ -135,7 +135,7 @@ export function ConnectionEnvironmentRow(props: {
         >
           {props.environment.isRelayManaged ? (
             <Text className="text-sm text-foreground-muted">
-              Managed by T3 Connect. Tunnel details update automatically.
+              Managed by T3 Cloud. Tunnel details update automatically.
             </Text>
           ) : (
             <>
