--- apps/web/src/components/settings/itemRows.ts
@@ -1,5 +1,0 @@
-/** Direct row in the card – same pattern as the Provider / ACP-agent list rows. */
-export const ITEM_ROW_CLASSNAME = "border-t border-border/60 px-4 py-4 first:border-t-0 sm:px-5";
-
-export const ITEM_ROW_INNER_CLASSNAME =
-  "flex flex-col gap-3 sm:flex-row sm:items-center sm:justify-between";
