+++ packages/client-runtime/src/relay/managedRelayState.ts
@@ -123,7 +123,7 @@ export function createManagedRelaySession(input: ManagedRelaySessionInput): Mana
         try: () => readCachedClerkToken(nowMillis),
         catch: (cause) =>
           new ManagedRelaySessionError({
-            message: "Could not obtain the T3 Connect session token.",
+            message: "Could not obtain the T3 Cloud session token.",
             cause,
           }),
       });
@@ -181,7 +181,7 @@ function readSessionClerkToken(
         ? Effect.succeed(token)
         : Effect.fail(
             new ManagedRelaySessionError({
-              message: "The T3 Connect session token is unavailable.",
+              message: "The T3 Cloud session token is unavailable.",
             }),
           ),
     ),
@@ -226,7 +226,7 @@ function requireClerkToken(
   if (!session || session.accountId !== accountId) {
     return Effect.fail(
       new ManagedRelaySessionError({
-        message: "Sign in to T3 Connect before loading relay data.",
+        message: "Sign in to T3 Cloud before loading relay data.",
       }),
     );
   }
@@ -296,7 +296,7 @@ export function readManagedRelaySnapshotState<A>(
   let errorTraceId: string | null = null;
   if (result._tag === "Failure") {
     const cause = Cause.squash(result.cause);
-    error = cause instanceof Error ? cause.message : "Could not load T3 Connect data.";
+    error = cause instanceof Error ? cause.message : "Could not load T3 Cloud data.";
     errorTraceId = findErrorTraceId(cause);
   }
   return {
